+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* initial release.
+
+## 0.0.2
+
+* add SubStateProviderBuilder and SubStateBuilder
+
+## 0.0.3
+
+* add final T Function(BuildContext context)create